@@ -63,9 +63,9 @@ type MsgSetSub struct {
 // MsgSetDesc is a C2S in set.what == "desc", acc, sub message.
 type MsgSetDesc struct {
 	DefaultAcs *MsgDefaultAcsMode `json:"defacs,omitempty"`  // default access mode
-	Public     interface{}        `json:"public,omitempty"`  // description of the user or topic
-	Trusted    interface{}        `json:"trusted,omitempty"` // trusted (system-provided) user or topic data
-	Private    interface{}        `json:"private,omitempty"` // per-subscription private data
+	Public     any                `json:"public,omitempty"`  // description of the user or topic
+	Trusted    any                `json:"trusted,omitempty"` // trusted (system-provided) user or topic data
+	Private    any                `json:"private,omitempty"` // per-subscription private data
 }
 
 // MsgCredClient is an account credential such as email or phone number.
@@ -77,7 +77,7 @@ type MsgCredClient struct {
 	// Verification response
 	Response string `json:"resp,omitempty"`
 	// Request parameters, such as preferences. Passed to valiator without interpretation.
-	Params map[string]interface{} `json:"params,omitempty"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // MsgSetQuery is an update to topic or user metadata: description, subscriptions, tags, credentials.
@@ -127,13 +127,14 @@ type MsgClientAcc struct {
 	Id string `json:"id,omitempty"`
 	// "newXYZ" to create a new user or UserId to update a user; default: current user.
 	User string `json:"user,omitempty"`
+	// Temporary authentication parameters for one-off actions, like password reset.
+	TmpScheme string `json:"tmpscheme,omitempty"`
+	TmpSecret []byte `json:"tmpsecret,omitempty"`
 	// Account state: normal, suspended.
 	State string `json:"status,omitempty"`
 	// Authentication level of the user when UserID is set and not equal to the current user.
 	// Either "", "auth" or "anon". Default: ""
 	AuthLevel string `json:"authlevel,omitempty"`
-	// Authentication token for resetting the password and maybe other one-time actions.
-	Token []byte `json:"token,omitempty"`
 	// The initial authentication scheme the account can use
 	Scheme string `json:"scheme,omitempty"`
 	// Shared secret
@@ -259,11 +260,11 @@ type MsgClientLeave struct {
 
 // MsgClientPub is client's request to publish data to topic subscribers {pub}.
 type MsgClientPub struct {
-	Id      string                 `json:"id,omitempty"`
-	Topic   string                 `json:"topic"`
-	NoEcho  bool                   `json:"noecho,omitempty"`
-	Head    map[string]interface{} `json:"head,omitempty"`
-	Content interface{}            `json:"content"`
+	Id      string         `json:"id,omitempty"`
+	Topic   string         `json:"topic"`
+	NoEcho  bool           `json:"noecho,omitempty"`
+	Head    map[string]any `json:"head,omitempty"`
+	Content any            `json:"content"`
 }
 
 // MsgClientGet is a query of topic state {get}.
@@ -446,11 +447,11 @@ type MsgTopicDesc struct {
 	ReadSeqId int `json:"read,omitempty"`
 	RecvSeqId int `json:"recv,omitempty"`
 	// Id of the last delete operation as seen by the requesting user
-	DelId   int         `json:"clear,omitempty"`
-	Public  interface{} `json:"public,omitempty"`
-	Trusted interface{} `json:"trusted,omitempty"`
+	DelId   int `json:"clear,omitempty"`
+	Public  any `json:"public,omitempty"`
+	Trusted any `json:"trusted,omitempty"`
 	// Per-subscription private data
-	Private interface{} `json:"private,omitempty"`
+	Private any `json:"private,omitempty"`
 }
 
 func (src *MsgTopicDesc) describe() string {
@@ -508,11 +509,11 @@ type MsgTopicSub struct {
 	// ID of the message reported by the given user as received
 	RecvSeqId int `json:"recv,omitempty"`
 	// Topic's public data
-	Public interface{} `json:"public,omitempty"`
+	Public any `json:"public,omitempty"`
 	// Topic's trusted public data
-	Trusted interface{} `json:"trusted,omitempty"`
+	Trusted any `json:"trusted,omitempty"`
 	// User's own private data per topic
-	Private interface{} `json:"private,omitempty"`
+	Private any `json:"private,omitempty"`
 
 	// Response to non-'me' topic
 
@@ -575,9 +576,9 @@ type MsgDelValues struct {
 
 // MsgServerCtrl is a server control message {ctrl}.
 type MsgServerCtrl struct {
-	Id     string      `json:"id,omitempty"`
-	Topic  string      `json:"topic,omitempty"`
-	Params interface{} `json:"params,omitempty"`
+	Id     string `json:"id,omitempty"`
+	Topic  string `json:"topic,omitempty"`
+	Params any    `json:"params,omitempty"`
 
 	Code      int       `json:"code"`
 	Text      string    `json:"text,omitempty"`
@@ -601,12 +602,12 @@ func (src *MsgServerCtrl) describe() string {
 type MsgServerData struct {
 	Topic string `json:"topic"`
 	// ID of the user who originated the message as {pub}, could be empty if sent by the system
-	From      string                 `json:"from,omitempty"`
-	Timestamp time.Time              `json:"ts"`
-	DeletedAt *time.Time             `json:"deleted,omitempty"`
-	SeqId     int                    `json:"seq"`
-	Head      map[string]interface{} `json:"head,omitempty"`
-	Content   interface{}            `json:"content"`
+	From      string         `json:"from,omitempty"`
+	Timestamp time.Time      `json:"ts"`
+	DeletedAt *time.Time     `json:"deleted,omitempty"`
+	SeqId     int            `json:"seq"`
+	Head      map[string]any `json:"head,omitempty"`
+	Content   any            `json:"content"`
 }
 
 // Deep-shallow copy.
@@ -907,13 +908,13 @@ func NoErrReply(msg *ClientComMessage, ts time.Time) *ServerComMessage {
 }
 
 // NoErrParams indicates successful completion with additional parameters (200).
-func NoErrParams(id, topic string, ts time.Time, params interface{}) *ServerComMessage {
+func NoErrParams(id, topic string, ts time.Time, params any) *ServerComMessage {
 	return NoErrParamsExplicitTs(id, topic, ts, ts, params)
 }
 
 // NoErrParamsExplicitTs indicates successful completion with additional parameters
 // and explicit server and incoming request timestamps (200).
-func NoErrParamsExplicitTs(id, topic string, serverTs, incomingReqTs time.Time, params interface{}) *ServerComMessage {
+func NoErrParamsExplicitTs(id, topic string, serverTs, incomingReqTs time.Time, params any) *ServerComMessage {
 	return &ServerComMessage{
 		Ctrl: &MsgServerCtrl{
 			Id:        id,
@@ -930,7 +931,7 @@ func NoErrParamsExplicitTs(id, topic string, serverTs, incomingReqTs time.Time, 
 
 // NoErrParamsReply indicates successful completion with additional parameters
 // and explicit server and incoming request timestamps (200).
-func NoErrParamsReply(msg *ClientComMessage, ts time.Time, params interface{}) *ServerComMessage {
+func NoErrParamsReply(msg *ClientComMessage, ts time.Time, params any) *ServerComMessage {
 	return NoErrParamsExplicitTs(msg.Id, msg.Original, ts, msg.Timestamp, params)
 }
 
@@ -970,7 +971,7 @@ func NoErrAcceptedExplicitTs(id, topic string, serverTs, incomingReqTs time.Time
 }
 
 // NoContentParams indicates request was processed but resulted in no content (204).
-func NoContentParams(id, topic string, serverTs, incomingReqTs time.Time, params interface{}) *ServerComMessage {
+func NoContentParams(id, topic string, serverTs, incomingReqTs time.Time, params any) *ServerComMessage {
 	return &ServerComMessage{
 		Ctrl: &MsgServerCtrl{
 			Id:        id,
@@ -987,7 +988,7 @@ func NoContentParams(id, topic string, serverTs, incomingReqTs time.Time, params
 
 // NoContentParamsReply indicates request was processed but resulted in no content
 // in response to a client request (204).
-func NoContentParamsReply(msg *ClientComMessage, ts time.Time, params interface{}) *ServerComMessage {
+func NoContentParamsReply(msg *ClientComMessage, ts time.Time, params any) *ServerComMessage {
 	return NoContentParams(msg.Id, msg.Original, ts, msg.Timestamp, params)
 }
 
@@ -1016,7 +1017,7 @@ func NoErrShutdown(ts time.Time) *ServerComMessage {
 }
 
 // NoErrDeliveredParams means requested content has been delivered (208).
-func NoErrDeliveredParams(id, topic string, ts time.Time, params interface{}) *ServerComMessage {
+func NoErrDeliveredParams(id, topic string, ts time.Time, params any) *ServerComMessage {
 	return &ServerComMessage{
 		Ctrl: &MsgServerCtrl{
 			Id:        id,
@@ -1059,7 +1060,7 @@ func InfoChallenge(id string, ts time.Time, challenge []byte) *ServerComMessage 
 			Id:        id,
 			Code:      http.StatusMultipleChoices, // 300
 			Text:      "challenge",
-			Params:    map[string]interface{}{"challenge": challenge},
+			Params:    map[string]any{"challenge": challenge},
 			Timestamp: ts,
 		},
 		Id:        id,
