@@ -63,7 +63,7 @@ func (sess *Session) readLoop() {
 	}
 }
 
-func (sess *Session) sendMessage(msg interface{}) bool {
+func (sess *Session) sendMessage(msg any) bool {
 	if len(sess.send) > sendQueueLimit {
 		logs.Err.Println("ws: outbound queue limit exceeded", sess.sid)
 		return false
@@ -144,7 +144,7 @@ func (sess *Session) writeLoop() {
 }
 
 // Writes a message with the given message type (mt) and payload.
-func wsWrite(ws *websocket.Conn, mt int, msg interface{}) error {
+func wsWrite(ws *websocket.Conn, mt int, msg any) error {
 	var bits []byte
 	if msg != nil {
 		bits = msg.([]byte)
@@ -157,8 +157,9 @@ func wsWrite(ws *websocket.Conn, mt int, msg interface{}) error {
 
 // Handles websocket requests from peers.
 var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
+	ReadBufferSize:    1024,
+	WriteBufferSize:   1024,
+	EnableCompression: globals.wsCompression,
 	// Allow connections from any Origin
 	CheckOrigin: func(r *http.Request) bool { return true },
 }
