@@ -36,7 +36,11 @@ func listenAndServe(addr string, mux *http.ServeMux, tlfConf *tls.Config, stop <
 	httpdone := make(chan bool)
 
 	server := &http.Server{
-		Handler: mux,
+		Handler:           mux,
+		ReadHeaderTimeout: 10 * time.Second,
+		IdleTimeout:       30 * time.Second,
+		WriteTimeout:      90 * time.Second,
+		MaxHeaderBytes:    1 << 14,
 	}
 
 	server.TLSConfig = tlfConf
@@ -398,13 +402,21 @@ type debugTopic struct {
 	Sessions []string `json:"sessions,omitempty"`
 }
 
+// debugCachedUser is a user cache entry debug info.
+type debugCachedUser struct {
+	Uid    string `json:"uid,omitempty"`
+	Unread int    `json:"unread,omitempty"`
+	Topics int    `json:"topics,omitempty"`
+}
+
 // debugDump is server internal state dump for debugging.
 type debugDump struct {
-	Version   string         `json:"server_version,omitempty"`
-	Build     string         `json:"build_id,omitempty"`
-	Timestamp time.Time      `json:"ts,omitempty"`
-	Sessions  []debugSession `json:"sessions,omitempty"`
-	Topics    []debugTopic   `json:"topics,omitempty"`
+	Version   string            `json:"server_version,omitempty"`
+	Build     string            `json:"build_id,omitempty"`
+	Timestamp time.Time         `json:"ts,omitempty"`
+	Sessions  []debugSession    `json:"sessions,omitempty"`
+	Topics    []debugTopic      `json:"topics,omitempty"`
+	UserCache []debugCachedUser `json:"user_cache,omitempty"`
 }
 
 func serveStatus(wrt http.ResponseWriter, req *http.Request) {
@@ -416,6 +428,7 @@ func serveStatus(wrt http.ResponseWriter, req *http.Request) {
 		Timestamp: types.TimeNow(),
 		Sessions:  make([]debugSession, 0, len(globals.sessionStore.sessCache)),
 		Topics:    make([]debugTopic, 0, 10),
+		UserCache: make([]debugCachedUser, 0, 10),
 	}
 	// Sessions.
 	globals.sessionStore.Range(func(sid string, s *Session) bool {
@@ -439,7 +452,7 @@ func serveStatus(wrt http.ResponseWriter, req *http.Request) {
 		return true
 	})
 	// Topics.
-	globals.hub.topics.Range(func(_, t interface{}) bool {
+	globals.hub.topics.Range(func(_, t any) bool {
 		topic := t.(*Topic)
 		psd := make([]string, 0, len(topic.sessions))
 		for s := range topic.sessions {
@@ -463,6 +476,13 @@ func serveStatus(wrt http.ResponseWriter, req *http.Request) {
 		})
 		return true
 	})
+	for k, v := range usersCache {
+		result.UserCache = append(result.UserCache, debugCachedUser{
+			Uid:    k.UserId(),
+			Unread: v.unread,
+			Topics: v.topics,
+		})
+	}
 
 	json.NewEncoder(wrt).Encode(result)
 }
