@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.21.6
+// - protoc-gen-go-grpc v1.3.0
+// - protoc             v3.21.12
 // source: model.proto
 
 package pbx
@@ -17,6 +17,10 @@ import (
 // is compatible with the grpc package it is being compiled against.
 // Requires gRPC-Go v1.32.0 or later.
 const _ = grpc.SupportPackageIsVersion7
+
+const (
+	Node_MessageLoop_FullMethodName = "/pbx.Node/MessageLoop"
+)
 
 // NodeClient is the client API for Node service.
 //
@@ -35,7 +39,7 @@ func NewNodeClient(cc grpc.ClientConnInterface) NodeClient {
 }
 
 func (c *nodeClient) MessageLoop(ctx context.Context, opts ...grpc.CallOption) (Node_MessageLoopClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Node_ServiceDesc.Streams[0], "/pbx.Node/MessageLoop", opts...)
+	stream, err := c.cc.NewStream(ctx, &Node_ServiceDesc.Streams[0], Node_MessageLoop_FullMethodName, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -138,6 +142,15 @@ var Node_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "model.proto",
 }
 
+const (
+	Plugin_FireHose_FullMethodName     = "/pbx.Plugin/FireHose"
+	Plugin_Find_FullMethodName         = "/pbx.Plugin/Find"
+	Plugin_Account_FullMethodName      = "/pbx.Plugin/Account"
+	Plugin_Topic_FullMethodName        = "/pbx.Plugin/Topic"
+	Plugin_Subscription_FullMethodName = "/pbx.Plugin/Subscription"
+	Plugin_Message_FullMethodName      = "/pbx.Plugin/Message"
+)
+
 // PluginClient is the client API for Plugin service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -171,7 +184,7 @@ func NewPluginClient(cc grpc.ClientConnInterface) PluginClient {
 
 func (c *pluginClient) FireHose(ctx context.Context, in *ClientReq, opts ...grpc.CallOption) (*ServerResp, error) {
 	out := new(ServerResp)
-	err := c.cc.Invoke(ctx, "/pbx.Plugin/FireHose", in, out, opts...)
+	err := c.cc.Invoke(ctx, Plugin_FireHose_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -180,7 +193,7 @@ func (c *pluginClient) FireHose(ctx context.Context, in *ClientReq, opts ...grpc
 
 func (c *pluginClient) Find(ctx context.Context, in *SearchQuery, opts ...grpc.CallOption) (*SearchFound, error) {
 	out := new(SearchFound)
-	err := c.cc.Invoke(ctx, "/pbx.Plugin/Find", in, out, opts...)
+	err := c.cc.Invoke(ctx, Plugin_Find_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -189,7 +202,7 @@ func (c *pluginClient) Find(ctx context.Context, in *SearchQuery, opts ...grpc.C
 
 func (c *pluginClient) Account(ctx context.Context, in *AccountEvent, opts ...grpc.CallOption) (*Unused, error) {
 	out := new(Unused)
-	err := c.cc.Invoke(ctx, "/pbx.Plugin/Account", in, out, opts...)
+	err := c.cc.Invoke(ctx, Plugin_Account_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -198,7 +211,7 @@ func (c *pluginClient) Account(ctx context.Context, in *AccountEvent, opts ...gr
 
 func (c *pluginClient) Topic(ctx context.Context, in *TopicEvent, opts ...grpc.CallOption) (*Unused, error) {
 	out := new(Unused)
-	err := c.cc.Invoke(ctx, "/pbx.Plugin/Topic", in, out, opts...)
+	err := c.cc.Invoke(ctx, Plugin_Topic_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -207,7 +220,7 @@ func (c *pluginClient) Topic(ctx context.Context, in *TopicEvent, opts ...grpc.C
 
 func (c *pluginClient) Subscription(ctx context.Context, in *SubscriptionEvent, opts ...grpc.CallOption) (*Unused, error) {
 	out := new(Unused)
-	err := c.cc.Invoke(ctx, "/pbx.Plugin/Subscription", in, out, opts...)
+	err := c.cc.Invoke(ctx, Plugin_Subscription_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -216,7 +229,7 @@ func (c *pluginClient) Subscription(ctx context.Context, in *SubscriptionEvent, 
 
 func (c *pluginClient) Message(ctx context.Context, in *MessageEvent, opts ...grpc.CallOption) (*Unused, error) {
 	out := new(Unused)
-	err := c.cc.Invoke(ctx, "/pbx.Plugin/Message", in, out, opts...)
+	err := c.cc.Invoke(ctx, Plugin_Message_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -292,7 +305,7 @@ func _Plugin_FireHose_Handler(srv interface{}, ctx context.Context, dec func(int
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pbx.Plugin/FireHose",
+		FullMethod: Plugin_FireHose_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PluginServer).FireHose(ctx, req.(*ClientReq))
@@ -310,7 +323,7 @@ func _Plugin_Find_Handler(srv interface{}, ctx context.Context, dec func(interfa
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pbx.Plugin/Find",
+		FullMethod: Plugin_Find_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PluginServer).Find(ctx, req.(*SearchQuery))
@@ -328,7 +341,7 @@ func _Plugin_Account_Handler(srv interface{}, ctx context.Context, dec func(inte
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pbx.Plugin/Account",
+		FullMethod: Plugin_Account_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PluginServer).Account(ctx, req.(*AccountEvent))
@@ -346,7 +359,7 @@ func _Plugin_Topic_Handler(srv interface{}, ctx context.Context, dec func(interf
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pbx.Plugin/Topic",
+		FullMethod: Plugin_Topic_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PluginServer).Topic(ctx, req.(*TopicEvent))
@@ -364,7 +377,7 @@ func _Plugin_Subscription_Handler(srv interface{}, ctx context.Context, dec func
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pbx.Plugin/Subscription",
+		FullMethod: Plugin_Subscription_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PluginServer).Subscription(ctx, req.(*SubscriptionEvent))
@@ -382,7 +395,7 @@ func _Plugin_Message_Handler(srv interface{}, ctx context.Context, dec func(inte
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pbx.Plugin/Message",
+		FullMethod: Plugin_Message_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PluginServer).Message(ctx, req.(*MessageEvent))
